src/pyproject.toml
[project]
name = "chat-app"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "bcrypt>=5.0.0",
    "fastapi[standard]>=0.119.1",
    "pydantic-settings>=2.11.0",
    "pyjwt>=2.10.1",
    "python-socketio>=5.14.3",
    "python-socketio-stubs>=5.14.1.1.20251014",
    "ruff>=0.14.1",
    "sqlalchemy[asyncio]>=2.0.44",
    "uvicorn[standard]>=0.38.0",
]

[dependency-groups]
dev = [
    "basedpyright>=1.32.0",
]

[tool.ruff]
target-version = "py314"
line-length = 88
indent-width = 4
fix = true
show-fixes = true

[tool.Ruff.lint]
# Comprehensive rule set for strict checking
select = [
    "E",      # pycodestyle errors
    "F",      # pyflakes
    "I",      # isort
    "UP",     # pyupgrade
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "T20",    # flake8-print (no print statements)
    "SIM",    # flake8-simplify
    "N",      # pep8-naming
    "Q",      # flake8-quotes
    "RUF",    # Ruff-specific rules
    "ASYNC",  # flake8-async
    "S",      # flake8-bandit (security)
    "PTH",    # flake8-use-pathlib
    "ERA",    # eradicate (commented-out code)
    "PL",     # pylint
    "PERF",   # perflint (performance)
    "ANN",    # flake8-annotations
    "ARG",    # flake8-unused-arguments
    "RET",    # flake8-return
    "TCH",    # flake8-type-checking
]

ignore = [
    "E501",    # Line too long (formatter handles this)
    "S603",    # subprocess without shell=True (too strict)
    "S607",    # Starting a process with a partial path (too strict)
    "ANN101",  # Missing type annotation for self (redundant)
    "ANN102",  # Missing type annotation for cls (redundant)
]

# Per-file ignores
[tool.ruff.lint.per-file-ignores]
"__init__.py" = [
    "F401",    # Allow unused imports in __init__.py
]
"tests/**/*.py" = [
    "S101",    # Allow assert in tests
    "PLR2004", # Allow magic values in tests
    "ANN",     # Don't require annotations in tests
]

[tool.ruff.lint.isort]
known-first-party = ["src"]  # CHANGE THIS
combine-as-imports = true
force-sort-within-sections = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-type-checking]
strict = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pyright]
pythonVersion = "3.14"
typeCheckingMode = "strict"

reportMissingImports = true
reportMissingTypeStubs = true  # Stricter: require type stubs
reportUndefinedVariable = true
reportAssertAlwaysTrue = true
reportInvalidStringEscapeSequence = true

reportOptionalSubscript = true
reportOptionalMemberAccess = true
reportOptionalCall = true
reportOptionalIterable = true
reportOptionalContextManager = true
reportOptionalOperand = true

reportMissingParameterType = true
reportMissingTypeArgument = true
reportUnknownParameterType = true
reportUnknownLambdaType = true
reportUnknownArgumentType = true   # STRICT: Enable (can be noisy)
reportUnknownVariableType = true   # STRICT: Enable (can be noisy)
reportUnknownMemberType = true     # STRICT: Enable (can be noisy)
reportUntypedFunctionDecorator = true
reportUntypedClassDecorator = true
reportUntypedBaseClass = true
reportUntypedNamedTuple = true

reportIncompatibleMethodOverride = true
reportIncompatibleVariableOverride = true
reportInconsistentConstructor = true
reportUninitializedInstanceVariable = true
reportOverlappingOverload = true
reportMissingSuperCall = true  # STRICT: Enable

reportPrivateUsage = true
reportConstantRedefinition = true
reportInvalidStubStatement = true
reportIncompleteStub = true
reportUnsupportedDunderAll = true
reportUnusedClass = "error"        # STRICT: Error instead of warning
reportUnusedFunction = "error"     # STRICT: Error instead of warning
reportUnusedVariable = "error"     # STRICT: Error instead of warning
reportUnusedImport = "error"       # STRICT: Error instead of warning
reportDuplicateImport = "error"    # STRICT: Error instead of warning

reportUnnecessaryIsInstance = "error"         # STRICT: Error
reportUnnecessaryCast = "error"               # STRICT: Error
reportUnnecessaryComparison = "error"         # STRICT: Error
reportUnnecessaryContains = "error"           # STRICT: Error
reportUnnecessaryTypeIgnoreComment = "error"  # STRICT: Error

reportGeneralTypeIssues = true
reportPropertyTypeMismatch = true
reportFunctionMemberAccess = true
reportCallInDefaultInitializer = true
reportImplicitStringConcatenation = true  # STRICT: Enable

reportImplicitOverride = true    # STRICT: Require @override decorator (Python 3.12+)
reportShadowedImports = true     # STRICT: Detect shadowed imports
reportDeprecated = "warning"     # Warn on deprecated usage
reportImportCycles = "warning"

exclude = [
    "**/__pycache__",
    "**/node_modules",
    ".git",
    ".mypy_cache",
    ".pyright_cache",
    ".ruff_cache",
    ".pytest_cache",
    ".venv",
    "venv",
    "env",
    "logs",
    "output",
    "data",
    "build",
    "dist",
    "*.egg-info",
]
